- [WIP/CLN] Splitting data classes into files
