+++ gempy/core/interpolator.py
@@ -1,6 +1,6 @@
 from typing import Union
-from gempy.core.data import Grid, Surfaces, AdditionalData
-from gempy.core.data_modules.geometric_data import SurfacePoints, Orientations
+from gempy.core.data import AdditionalData
+from gempy import SurfacePoints, Orientations, Surfaces, Grid
 from gempy.core.data_modules.stack import Faults, Series
 from gempy.utils.meta import _setdoc_pro, _setdoc
 import gempy.utils.docstring as ds
